personapi/requirements.txt
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
dill==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.3.5
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
